Add files via upload
We established key business objectives and success metrics to evaluate performance based on the company’s characteristics and public statements about its goals. We then accessed the Google Merchandise Store account using Google Analytics and assumed the data available
is representative of our company i.e. Nordstorm. 
We used Oct 2020 as timeline ( vs Oc 2019) to provide a summary of the company’s eCommerce performance and analyzed the factors impacting the selected KPIs. Based on the findings of our analysis, we provided actionable commendations for business improvement.